Added buttons to the Details Page and a refresh button on the fravourites page. Altered how the calendar is displayed and closed
Made the calendar abit more user-friendly with automated closing. Before
it was not closing as you switch between weather locations. Added the
previous and forecast functions to the Details page and edited the
loading gif to have a transparent background. Also date is now showing
on info table when calendar date is selected